# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: |
    Namma Seva API (Phase 3). The chain is the source of truth; this API serves the indexed
    Postgres read model, pins metadata to IPFS for wallets to sign, authenticates wallets (SIWE)
    and citizens (phone OTP), and relays gasless citizen grievances through the ERC-2771 forwarder.
    Officials, auditors and contractors sign privileged transactions in their own wallets — there
    are no server-signed approve / reject / release routes.

    Amounts are decimal strings in the escrow's accounting unit (INR paise in LEDGER mode, wei in
    ESCROW mode). Coordinates are micro-degrees (`latE6`, `lngE6`).
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        details: {}
      required: [error]
...
    Bytes32:
      type: string
      pattern: "^0x[0-9a-fA-F]{64}$"
...
    Amount:
      type: string
      pattern: "^[0-9]+$"
      description: Non-negative integer as a decimal string.
...
    ProjectStatus:
      type: string
      enum: [PENDING_APPROVAL, ACTIVE, PAUSED, COMPLETED, CANCELLED]
...
    MilestoneStatus:
      type: string
      enum: [PENDING, PROOF_SUBMITTED, APPROVED, REJECTED, PAID, VOID]
    Milestone:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        title: { type: [string, "null"] }
        description: { type: [string, "null"] }
        amount: { $ref: "#/components/schemas/Amount" }
        status: { $ref: "#/components/schemas/MilestoneStatus" }
        round: { type: integer }
        approvalCount: { type: integer }
        metaCid: { type: string }
        proofCid: { type: [string, "null"] }
        proofHash: { type: [string, "null"] }
        proofLatE6: { type: [integer, "null"] }
        proofLngE6: { type: [integer, "null"] }
        submittedBy: { type: [string, "null"] }
        submittedAt: { type: [string, "null"], format: date-time }
        paidTx: { type: [string, "null"] }
        paidAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        createdTx: { type: string }
        updatedBlock: { type: integer }
      required:
        [id, projectId, title, description, amount, status, round, approvalCount, metaCid, proofCid, proofHash,
         proofLatE6, proofLngE6, submittedBy, submittedAt, paidTx, paidAt, createdAt, createdTx, updatedBlock]
    ProofMedia:
      type: object
      properties:
        cid: { type: string }
        sha256: { type: string }
        mime: { type: string }
        exifLat: { type: [number, "null"] }
        exifLng: { type: [number, "null"] }
        exifTime: { type: [string, "null"], format: date-time }
        gpsDistanceM: { type: [number, "null"] }
        flagged: { type: boolean }
        thumbCid: { type: [string, "null"] }
        checks:
          type: [object, "null"]
          description: Integrity checks (`capture`, `geofence`, `time`, `duplicate`). A failed check flags the photo but never blocks submission.
          additionalProperties:
            type: object
            properties:
              ok: { type: boolean }
              detail: { type: string }
            required: [ok]
      required: [cid, sha256, mime, exifLat, exifLng, exifTime, gpsDistanceM, flagged, thumbCid, checks]
    ProofUploadForm:
      type: object
      properties:
        photos:
          type: array
          maxItems: 5
          items: { type: string, format: binary }
        latE6: { type: integer, description: Device GPS fix, used when a photo has no EXIF GPS. }
        lngE6: { type: integer }
        note: { type: string, maxLength: 1000 }
        source:
          type: string
          enum: [camera, gallery]
          description: How the photo was captured. In strict mode anything but `camera` fails the capture check.
      required: [photos]
    ProofUploadResponse:
      type: object
      description: Arguments for `MilestoneEscrow.submitProof(milestoneId, proofCID, proofHash, latE6, lngE6)`.
      properties:
        milestoneId: { type: integer }
        proofCID: { type: string }
        proofHash: { $ref: "#/components/schemas/Bytes32" }
        latE6: { type: integer }
        lngE6: { type: integer }
        media:
          type: array
          items: { $ref: "#/components/schemas/ProofMedia" }
        warnings:
          type: array
          items: { type: string }
      required: [milestoneId, proofCID, proofHash, latE6, lngE6, media, warnings]

    # ─── Grievances ─────────────────────────────────────────────────────
...
    MilestoneWithProject:
      allOf:
        - $ref: "#/components/schemas/Milestone"
        - type: object
          properties:
            approvers:
              type: array
              description: Auditors who approved in the current review round.
              items: { type: string }
            project:
              type: object
              properties:
                id: { type: integer }
                title: { type: [string, "null"] }
                status: { $ref: "#/components/schemas/ProjectStatus" }
                wardId: { type: integer }
                latE6: { type: integer }
                lngE6: { type: integer }
                officialAddr: { type: string }
                contractorAddr: { type: [string, "null"] }
                approvalThreshold: { type: integer }
              required: [id, title, status, wardId, latE6, lngE6, officialAddr, contractorAddr, approvalThreshold]
          required: [approvers, project]
...
security: []
paths:
...
  /milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: Milestones with their project and current-round approvers (dashboard queues)
      parameters:
        - { name: projectId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/MilestoneStatus" } }
        - { name: contractor, in: query, schema: { type: string } }
        - { name: official, in: query, schema: { type: string } }
        - { name: wardId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MilestoneWithProject" }
  /milestones/pending:
    get:
      operationId: listPendingMilestones
      tags: [milestones]
      summary: Milestones with proof awaiting auditor review
      parameters:
        - { name: wardId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Pending review
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Milestone" }
  /milestones/{id}:
    get:
      operationId: getMilestone
      tags: [milestones]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Milestone
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Milestone" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /milestones/{id}/proof/upload:
    post:
      operationId: uploadMilestoneProof
      tags: [milestones]
      security: [{ bearer: [] }]
      summary: Upload geo-tagged photos; returns `proofCID` + `proofHash` for `submitProof`
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/ProofUploadForm" }
      responses:
        "201":
          description: Proof pinned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProofUploadResponse" }
        "400":
          description: Invalid upload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /milestones/{id}/proof:
    get:
      operationId: getMilestoneProof
      tags: [milestones]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Uploaded proof media for the milestone
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProofMedia" }

  # ─── Grievances ────────────────────────────────────────────────────────
...